pyproject.toml
[project]
name = "llm-council"
version = "0.1.0"
description = "Your LLM Council"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.9.0",
    "email-validator>=2.0.0",
    "tenacity>=8.2.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "greenlet>=3.0.0",
    "alembic>=1.12.0",
    "python-jose>=3.4.0",
    "passlib>=1.7.4",
    "bcrypt>=4.2.0",
    "redis>=5.0.0",
    "python-multipart",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0,<9",
    "pytest-asyncio>=0.21.0,<1",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "pytest-postgresql>=5.0.0",
    "aiosqlite>=0.19.0,<0.20",
    "fakeredis>=2.20.0",
]
